+++ legacy/expressions/migrate.go
@@ -15,6 +15,9 @@ import (
 	"github.com/antlr/antlr4/runtime/Go/antlr"
 )
 
+// ContextTopLevels are the allowed top-level identifiers in legacy expressions, i.e. @contact.bar is valid but @foo.bar isn't
+var ContextTopLevels = []string{"channel", "child", "contact", "date", "extra", "flow", "parent", "step"}
+
 var functionReturnTypes = map[string]string{
 	"abs":                 "number",
 	"datetime_add":        "datetime",
